+++ src/components/textfield/index.tsx
@@ -77,7 +77,10 @@ const TextField = (props: TextFieldTypeProps) => {
             ? 'unset'
             : `calc(14.5px - ${varHeight}px)`,
           flex: '1 0 0',
-          color: props.value ? 'var(--black)' : 'var(--accent-400)',
+          color:
+            props.value || props.inputProps?.value
+              ? 'var(--black)'
+              : 'var(--accent-400)',
           cursor: props.disabled && 'not-allowed',
         },
         '.MuiInput-root:before': {
@@ -144,7 +147,7 @@ const TextField = (props: TextFieldTypeProps) => {
 
         '.MuiFormLabel-root[data-shrink=false]': { top: `-${varHeight}px` },
         '& > .MuiAutocomplete-popupIndicator': {
-          '& svg, & svg g, & svg g path': 'var(--black)',
+          '& svg, & svg g, & svg g path': { fill: 'var(--black)' },
         },
 
         '& .MuiInputAdornment-positionStart .MuiSvgIcon-root': {
@@ -155,6 +158,9 @@ const TextField = (props: TextFieldTypeProps) => {
           color: !props.disabled
             ? endIcon?.props.color || 'var(--black)'
             : 'var(--accent-200)',
+          '& g path': {
+            fill: endIcon?.props.color || 'var(--black)',
+          },
         },
         ...props.sx,
       }}