+++ src/features/ministry/hooks/useMinistryMonthlyRecord.tsx
@@ -75,8 +75,12 @@ const useMinistryMonthlyRecord = ({
   }, [delegatedReports, month, person_uid]);
 
   const status = useMemo(() => {
-    if (congReport) {
-      return congReport.report_data.status;
+    if (!publisher) {
+      return congReport?.report_data.status || '';
+    }
+
+    if (congReport?.report_data.status === 'confirmed') {
+      return 'confirmed';
     }
 
     if (isSelf) {
@@ -88,10 +92,14 @@ const useMinistryMonthlyRecord = ({
     if (!delegatedReport) return 'pending';
 
     return delegatedReport.report_data.status;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const shared_ministry = useMemo(() => {
-    if (congReport) {
+    if (!publisher) {
+      return congReport?.report_data.shared_ministry ?? false;
+    }
+
+    if (congReport?.report_data.status === 'confirmed') {
       return congReport.report_data.shared_ministry;
     }
 
@@ -104,10 +112,14 @@ const useMinistryMonthlyRecord = ({
     if (!delegatedReport) return false;
 
     return delegatedReport.report_data.shared_ministry;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const comments = useMemo(() => {
-    if (congReport) {
+    if (!publisher) {
+      return congReport?.report_data.comments || '';
+    }
+
+    if (congReport?.report_data.status === 'confirmed') {
       return congReport.report_data.comments;
     }
 
@@ -120,10 +132,14 @@ const useMinistryMonthlyRecord = ({
     if (!delegatedReport) return '';
 
     return delegatedReport.report_data.comments;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const bible_studies = useMemo(() => {
-    if (congReport) {
+    if (!publisher) {
+      return congReport?.report_data.bible_studies ?? 0;
+    }
+
+    if (congReport?.report_data.status === 'confirmed') {
       return congReport.report_data.bible_studies;
     }
 
@@ -151,7 +167,7 @@ const useMinistryMonthlyRecord = ({
     const totalCount = delegatedReport.report_data.bible_studies.monthly;
 
     return totalCount < recordsCount ? recordsCount : totalCount;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const bible_studies_records = useMemo(() => {
     if (!isSelf) return;
@@ -180,11 +196,15 @@ const useMinistryMonthlyRecord = ({
     if (isInactive) return true;
 
     if (publisher) {
-      return congReports.some(
+      const congReport = congReports.find(
         (record) =>
           record.report_data.report_date === month &&
           record.report_data.person_uid === person_uid
       );
+
+      if (!congReport) return false;
+
+      return congReport.report_data.status === 'confirmed';
     }
 
     if (!branchReport) return false;
@@ -211,6 +231,10 @@ const useMinistryMonthlyRecord = ({
       return `${congReport.report_data.hours.field_service}:00`;
     }
 
+    if (!publisher) {
+      return '0:00';
+    }
+
     if (isSelf) {
       if (!userReport) return '0:00';
 
@@ -218,13 +242,12 @@ const useMinistryMonthlyRecord = ({
         return `${userReport.report_data.hours.field_service}:00`;
       }
 
-      const [hoursDaily, minutesDaily] =
-        userReport.report_data.hours.field_service.daily.split(':').map(Number);
+      const daily = userReport.report_data.hours.field_service.daily || '0:00';
+      const monthly =
+        userReport.report_data.hours.field_service.monthly || '0:00';
 
-      const [hoursMonthly, minutesMonthly] =
-        userReport.report_data.hours.field_service.monthly
-          .split(':')
-          .map(Number);
+      const [hoursDaily, minutesDaily] = daily.split(':').map(Number);
+      const [hoursMonthly, minutesMonthly] = monthly.split(':').map(Number);
 
       let totalHours = hoursDaily + hoursMonthly;
       const totalMinutes = (minutesDaily || 0) + (minutesMonthly || 0);
@@ -238,11 +261,15 @@ const useMinistryMonthlyRecord = ({
     if (!delegatedReport) return '0:00';
 
     return delegatedReport.report_data.hours.field_service.monthly;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const hours_credits = useMemo(() => {
     if (congReport) {
-      return `${congReport.report_data.hours.credit.approved}`;
+      return `${congReport.report_data.hours.credit.approved}:00`;
+    }
+
+    if (!publisher) {
+      return '0:00';
     }
 
     if (isSelf) {
@@ -255,11 +282,11 @@ const useMinistryMonthlyRecord = ({
         return `${userReport.report_data.hours.credit['value']}:00`;
       }
 
-      const [hoursDaily, minutesDaily] =
-        userReport.report_data.hours.credit.daily.split(':').map(Number);
+      const daily = userReport.report_data.hours.credit.daily || '0:00';
+      const monthly = userReport.report_data.hours.credit.monthly || '0:00';
 
-      const [hoursMonthly, minutesMonthly] =
-        userReport.report_data.hours.credit.monthly.split(':').map(Number);
+      const [hoursDaily, minutesDaily] = daily.split(':').map(Number);
+      const [hoursMonthly, minutesMonthly] = monthly.split(':').map(Number);
 
       let totalHours = hoursDaily + hoursMonthly;
       const totalMinutes = (minutesDaily || 0) + (minutesMonthly || 0);
@@ -273,13 +300,17 @@ const useMinistryMonthlyRecord = ({
     if (!delegatedReport) return '0:00';
 
     return delegatedReport.report_data.hours.credit.monthly;
-  }, [congReport, userReport, isSelf, delegatedReport]);
+  }, [publisher, congReport, userReport, isSelf, delegatedReport]);
 
   const hours_total = useMemo(() => {
     if (congReport) {
       return String(
         +hours_fields.split(':').at(0) + +hours_credits.split(':').at(0)
       );
+    }
+
+    if (!publisher) {
+      return '0';
     }
 
     const [hoursDaily, minutesDaily] = hours_fields.split(':').map(Number);
@@ -293,7 +324,7 @@ const useMinistryMonthlyRecord = ({
     totalHours += (totalMinutes - minutesRemain) / 60;
 
     return `${totalHours}:${String(minutesRemain).padStart(2, '0')}`;
-  }, [congReport, hours_fields, hours_credits]);
+  }, [publisher, congReport, hours_fields, hours_credits]);
 
   const minutes_remains = useMemo(() => {
     const minutes = hours_total.split(':').at(1);