+++ src/features/reports/meeting_attendance/monthly_record/week_box/useWeekBox.tsx
@@ -1,20 +1,35 @@
 import { ChangeEvent, useEffect, useMemo, useState } from 'react';
 import { useRecoilValue } from 'recoil';
+import { useAppTranslation } from '@hooks/index';
 import { formatDate } from '@services/dateformat';
-import { getWeekDate, weeksInMonth } from '@utils/date';
+import {
+  addDays,
+  addWeeks,
+  firstWeekMonth,
+  getWeekDate,
+  weeksInMonth,
+} from '@utils/date';
 import { WeekBoxProps } from './index.types';
 import { meetingAttendanceState } from '@states/meeting_attendance';
 import { WeeklyAttendance } from '@definition/meeting_attendance';
 import {
   attendanceOnlineRecordState,
+  midweekMeetingWeekdayState,
   userDataViewState,
+  weekendMeetingWeekdayState,
 } from '@states/settings';
 import { meetingAttendancePresentSave } from '@services/app/meeting_attendance';
+import { monthShortNamesState } from '@states/app';
 
 const useWeekBox = ({ month, index, type }: WeekBoxProps) => {
+  const { t } = useAppTranslation();
+
   const attendances = useRecoilValue(meetingAttendanceState);
   const dataView = useRecoilValue(userDataViewState);
   const recordOnline = useRecoilValue(attendanceOnlineRecordState);
+  const midweekDay = useRecoilValue(midweekMeetingWeekdayState);
+  const weekendDay = useRecoilValue(weekendMeetingWeekdayState);
+  const months = useRecoilValue(monthShortNamesState);
 
   const [present, setPresent] = useState('');
   const [online, setOnline] = useState('');
@@ -100,6 +115,34 @@ const useWeekBox = ({ month, index, type }: WeekBoxProps) => {
     }
   }, [weeksList, index, type, isMidweek, isWeekend]);
 
+  const box_label = useMemo(() => {
+    const [year, monthValue] = month.split('/').map(Number);
+
+    const firstWeek = firstWeekMonth(year, monthValue);
+
+    let meetingDate: Date;
+
+    if (type === 'midweek') {
+      meetingDate = addDays(firstWeek, midweekDay - 1);
+    }
+
+    if (type === 'weekend') {
+      meetingDate = addDays(firstWeek, weekendDay - 1);
+    }
+
+    meetingDate = addWeeks(meetingDate, index - 1);
+
+    const monthIndex = meetingDate.getMonth();
+    const date = meetingDate.getDate();
+
+    const dateLabel = t('tr_longDateNoYearLocale', {
+      month: months[monthIndex],
+      date,
+    });
+
+    return dateLabel;
+  }, [month, type, midweekDay, weekendDay, index, t, months]);
+
   const handlePresentChange = (e: ChangeEvent<HTMLInputElement>) => {
     if (e.target.value.match(/\D/)) {
       e.preventDefault();
@@ -150,6 +193,7 @@ const useWeekBox = ({ month, index, type }: WeekBoxProps) => {
     total,
     isMidweek,
     isWeekend,
+    box_label,
   };
 };