+++ src/pages/congregation/manage_access/all_users/useAllUsers.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { apiCongregationUsersGet } from '@services/api/congregation';
@@ -8,7 +8,6 @@ import {
   congregationsPersonsState,
   congregationUsersState,
 } from '@states/app';
-import { settingsState } from '@states/settings';
 
 const useAllUsers = () => {
   const { data, isLoading } = useQuery({
@@ -22,36 +21,12 @@ const useAllUsers = () => {
   const congregationsPersons = useRecoilValue(congregationsPersonsState);
   const appAdmin = useRecoilValue(congregationsAppAdminState);
   const baptizedPersons = useRecoilValue(congregationsBaptizedPersonsState);
-  const settings = useRecoilValue(settingsState);
-
-  const sync_enabled = useMemo(() => {
-    return settings.cong_settings.data_sync.value;
-  }, [settings.cong_settings.data_sync.value]);
 
   const [userAddOpen, setUserAddOpen] = useState(false);
-  const [enableSyncOpen, setEnableSyncOpen] = useState(false);
+
+  const handleOpenUserAdd = () => setUserAddOpen(true);
 
   const handleCloseUserAdd = () => setUserAddOpen(false);
-
-  const handleCloseEnableSync = () => setEnableSyncOpen(false);
-
-  const handleAction = () => {
-    if (sync_enabled) {
-      setUserAddOpen(true);
-    }
-
-    if (!sync_enabled) {
-      setEnableSyncOpen(true);
-    }
-  };
-
-  const handleContinue = () => {
-    setEnableSyncOpen(false);
-
-    if (sync_enabled) {
-      setUserAddOpen(true);
-    }
-  };
 
   useEffect(() => {
     if (Array.isArray(data?.users)) {
@@ -65,11 +40,8 @@ const useAllUsers = () => {
     congregationsPersons,
     baptizedPersons,
     appAdmin,
-    handleAction,
-    handleContinue,
-    enableSyncOpen,
-    handleCloseEnableSync,
     isLoading,
+    handleOpenUserAdd,
   };
 };