+++ src/views/reports/S21/shared/Header.tsx
@@ -1,13 +1,14 @@
 import { Text, View } from '@react-pdf/renderer';
 import { useAppTranslation } from '@hooks/index';
 import { styles } from './index.styles';
+import { HeaderProps } from './index.types';
 
-const Header = () => {
+const Header = ({ lang }: HeaderProps) => {
   const { t } = useAppTranslation();
 
   return (
     <View style={styles.header}>
-      <Text>{t('tr_S21Header').toUpperCase()}</Text>
+      <Text>{t('tr_S21Header', { lng: lang }).toUpperCase()}</Text>
     </View>
   );
 };